+++ cookbooks/build-essential/recipes/default.rb
@@ -21,13 +21,14 @@ require 'chef/shell_out'
 
 compiletime = node['build_essential']['compiletime']
 
-case node['os']
-when "linux"
+case node['platform_family']
+when "rhel", "suse", "fedora", "debian"
 
   # on apt-based platforms when first provisioning we need to force
   # apt-get update at compiletime if we are going to try to install at compiletime
   if node['platform_family'] == "debian"
-    execute "apt-get update" do
+    execute "apt-get-update-build-essentials" do
+      command "apt-get update"
       action :nothing
       # tip: to suppress this running every time, just use the apt cookbook
       not_if do
@@ -42,6 +43,8 @@ when "linux"
       %w{build-essential binutils-doc}
     when "rhel", "fedora"
       %w{gcc gcc-c++ kernel-devel make}
+    when "suse"
+      %w{gcc gcc-c++ kernel-default-devel make m4}  # in SLES there is no kernel-devel
     end
 
   packages.each do |pkg|
@@ -57,9 +60,19 @@ when "linux"
     end
     r.run_action(:install) if compiletime
   end
-when "darwin"
+
+when "smartos"
+    include_recipe 'pkgin'
+    %w{gcc47 gcc47-runtime scmgit-base gmake pkg-config binutils}.each do |package|
+      pkgin_package package do
+        action :install
+      end
+    end
+
+when "mac_os_x"
   result = Chef::ShellOut.new("pkgutil --pkgs").run_command
-  installed = result.stdout.split("\n").include?("com.apple.pkg.gcc4.2Leo")
+  osx_gcc_installer_installed = result.stdout.split("\n").include?("com.apple.pkg.gcc4.2Leo")
+  developer_tools_cli_installed = result.stdout.split("\n").include?("com.apple.pkg.DeveloperToolsCLI")
   pkg_filename = File.basename(node['build_essential']['osx']['gcc_installer_url'])
   pkg_path = "#{Chef::Config[:file_cache_path]}/#{pkg_filename}"
 
@@ -67,13 +80,13 @@ when "darwin"
     source node['build_essential']['osx']['gcc_installer_url']
     checksum node['build_essential']['osx']['gcc_installer_checksum']
     action ( compiletime ? :nothing : :create )
-    not_if { installed }
+    not_if { osx_gcc_installer_installed or developer_tools_cli_installed  }
   end
   r.run_action(:create) if compiletime
 
   r = execute "sudo installer -pkg \"#{pkg_path}\" -target /" do
     action ( compiletime ? :nothing : :run )
-    not_if { installed }
+    not_if { osx_gcc_installer_installed or developer_tools_cli_installed  }
   end
   r.run_action(:run) if compiletime
 end