+++ cookbooks/build-essential/CHANGELOG.md
@@ -1,3 +1,23 @@
+## v1.3.4:
+
+* [COOK-2272] - Complete `platform_family` conversion in build-essential
+
+## v1.3.2:
+
+* [COOK-2069] - build-essential will install osx-gcc-installer when
+  Xcode is present
+
+## v1.3.0:
+
+ * [COOK-1895] - support smartos
+
+## v1.2.0:
+
+* Add test-kitchen support (source repo only)
+* [COOK-1677] - build-essential cookbook support for OpenSuse and SLES
+* [COOK-1718] - build-essential cookbook metadata should include scientific
+* [COOK-1768] - The apt-get update in build-essentials needs to be renamed
+
 ## v1.1.2:
 
 * [COOK-1620] - support OS X 10.8